requirements.txt
bleach==3.2.1
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
Pillow==8.0.1
python-dotenv==0.15.0
python-slugify==4.0.1
